+++ test/integration_tests/cli/diff_command_test.dart
@@ -143,5 +143,25 @@ void main() {
       },
       timeout: integrationTestTimeout,
     );
+
+    test(
+      'diffing cloud_firestore 4.3.1 to latest works',
+      () async {
+        // just some random package for testing the diff command for pub refs without version
+        final diffCommand = DiffCommand();
+        final runner =
+            CommandRunner<int>('dart_apitool_tests', 'Test for dart_apitool')
+              ..addCommand(diffCommand);
+        final exitCode = await runner.run([
+          'diff',
+          '--old',
+          'pub://cloud_firestore/4.3.1',
+          '--new',
+          'pub://cloud_firestore',
+        ]);
+        expect(exitCode, 0);
+      },
+      timeout: integrationTestTimeout,
+    );
   });
 }